backend/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.1.0
bidict==0.22.0
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.2.1
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.1
opencv-python==4.6.0.66
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-engineio==4.3.4
python-socketio==5.7.1
python-telegram-bot==13.13
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rfc3986==1.5.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
timeloop==1.0.2
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
Werkzeug==2.2.1
yarl==1.8.1
